Infer mode for already running process
Infer mode for popen() from the process instead of specifying it
redundantly.  The process must be prepared so that exactly one of stdin
and stdout, and never stderr, is an open pipe.  If one of the other
streams is opened as a pipe initially, it must be closed or handed off to
a thread and set to None in the process object before calling popen().