Use correct mode for subprocess pipes
Use 'rb' and 'wb' modes for pipes to subprocess since that what they
actually are.  Binary and text modes must be clearly distinguished in
Python 3.

For now each line is decoded explicitly when text is required.  Ideally,
files should be opened in text mode or wrapped with TextIOWrapper to get
text output.  This is tricky because subprocess opens stdout and stderr
with the same mode, and in Python 2 returns a binary stream that is not
compatible with TextIOWrapper.